requirements.txt
torch>=2.1
torchvision>=0.16
numpy
pyyaml
tqdm
matplotlib
lensless
git+https://github.com/ebezzam/slm-controller.git
waveprop
git+https://github.com/pvigier/perlin-numpy.git@5e26837db14042e51166eb6cad4c0df2c1907016
datasets
wandb